# backend/srcs/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST du projet Pan Bagnat.",
        "title": "Pan Bagnat API",
        "contact": {},
        "version": "1.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/module-page/{pageName}": {
            "get": {
                "description": "Reverse-proxies /module-page/{pageName} to the module’s configured URL.",
                "tags": [
                    "Pages"
                ],
                "summary": "Proxy Module Page (root)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the module page",
                        "name": "pageName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proxied content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Module page name not provided",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error looking up or proxying the module page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/module-page/{pageName}/{path}": {
            "get": {
                "description": "Reverse-proxies /module-page/{pageName}/{path}/* to the module’s configured URL, stripping the prefix.",
                "tags": [
                    "Pages"
                ],
                "summary": "Proxy Module Page (sub-paths)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the module page",
                        "name": "pageName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sub-path under the module page (may include slashes)",
                        "name": "path",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proxied content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Module page name or path not provided",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error looking up or proxying the module page",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/modules": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns all available modules for your campus, with optional filtering, sorting, and pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Get Module List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter expression (e.g. \\",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination token for the next page",
                        "name": "next_page_token",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order: asc or desc",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.ModuleGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Downloads and registers a new module for your campus by cloning from the provided Git URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Post Module",
                "parameters": [
                    {
                        "description": "Git URL, branch, and module name",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.ModuleGitInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly imported module",
                        "schema": {
                            "$ref": "#/definitions/api.Module"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to import module",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Delete a module for your campus (All module datas will be lost!)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Delete Module",
                "parameters": [
                    {
                        "description": "Module input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.ModulePatchInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/modules/{moduleID}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns all information about a module given its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Get Module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Module"
                        }
                    },
                    "400": {
                        "description": "ID not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the specified fields of a module (e.g. name, Git URL/branch, status).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Patch Module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.ModulePatchInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated module",
                        "schema": {
                            "$ref": "#/definitions/api.Module"
                        }
                    },
                    "400": {
                        "description": "Invalid module ID or JSON body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Module not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/modules/{moduleID}/docker/config": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Returns the module’s config.yml as a YAML string under the `config` field.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules",
                    "Docker"
                ],
                "summary": "Get Module Configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module ID",
                        "name": "moduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.ConfigResponse"
                        }
                    },
                    "400": {
                        "description": "ID not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "modules.ConfigResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "Config is the raw module.yml content, with newlines preserved",
                    "type": "string",
                    "example": "foo: bar\nbaz: qux\n"
                }
            }
        },
        "modules.ModuleGetResponse": {
            "type": "object",
            "properties": {
                "modules": {
                    "description": "Modules is the list of modules on this page.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Module"
                    }
                },
                "next_page_token": {
                    "description": "NextPageToken is the token to retrieve the next page of results.",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
            }
        },
        "modules.ModuleGitInput": {
            "type": "object",
            "properties": {
                "git_branch": {
                    "description": "GitBranch is the branch to check out. Defaults to \"main\" if omitted.",
                    "type": "string",
                    "example": "main"
                },
                "git_url": {
                    "description": "GitURL is the repository URL to clone the module from.",
                    "type": "string",
                    "example": "https://github.com/some-user/some-repo"
                },
                "name": {
                    "description": "Name is the human-readable title you want for this module.",
                    "type": "string",
                    "example": "Captain Hook"
                }
            }
        },
...
        "modules.ModulePatchInput": {
            "description": "API Module model",
            "type": "object",
            "properties": {
                "gitBranch": {
                    "type": "string",
                    "example": "main"
                },
                "git_url": {
                    "type": "string",
                    "example": "https://github.com/some-user/some-repo"
                },
                "name": {
                    "type": "string",
                    "example": "captain-hook"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    },
...
}